+++ CHANGELOG.md
@@ -26,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Effects are now bound via internal mechanism during CreateEffect.
 - Simplified Example folder structures.
 - Many tests now directly use existing Fluxity type rather than a dummy.
-- Feature<TState>s are no longer added to the ServiceInstaller. To get direct access to a feature either request the `Store` and `GetFeatureView` or use a `FeatureView`.
+- Feature<TState>s are no longer added to the ServiceInstaller. To get direct access to a feature either request the `Store` and `GetFeatureView` or use a `FeatureObserver`.
 
 ### Removed